+++ src/main/java/com/pearson/app/dao/ImageDAOInterface.java
@@ -9,6 +9,6 @@ public interface ImageDAOInterface {
     
     public List<Image> list();
     public Image create(Image image);
-    public Image get(Long id);
+    public Image get(int id);
     public void delete(Image image);
 }
