+++ src/main/java/com/pearson/app/dao/UserDAOInterface.java
@@ -12,7 +12,7 @@ public interface UserDAOInterface {
 
         public void addUser(User user);
         public List<User> listUsers();
-        public User getUserById(Long id);
+        public User getUserById(Integer id);
         public User getUserByUsername(String username);
         public void updateUser(User user);
         public void removeUser(Long id);
