+++ src/main/java/com/pearson/app/services/ImageService.java
@@ -38,7 +38,7 @@ public class ImageService {
     }
 
     @Transactional(readOnly = true)
-    public Image getImage(Long id) {
+    public Image getImage(int id) {
         return imageRepository.get(id);
     }
