+++ src/main/java/com/pearson/app/services/TransformationServiceInterface.java
@@ -14,7 +14,7 @@ public interface TransformationServiceInterface {
 
         public void addTransformation(Transformation transformation);
         public List<Transformation> listTransformations();
-        public Transformation getTransformationById(Long id);
+        public Transformation getTransformationById(int id);
         public void updateTransformation(Transformation transformation);
-        public void removeTransformation(Long id);
+        public void removeTransformation(int id);
 }
